refactor(settings): 移动配置文件读写及默认常量至settings模块
- 将默认模型和基础URL常量移至settings模块，统一管理
- 实现用户配置和项目配置的读取、写入接口
- 提供获取用户和项目配置路径的工具函数
- 清理ui/utils中的相关冗余代码，简化代码结构
- 统一类型引用，调整session、session-types、settings和common之间的导入路径
- 移除AsciiArt模块内容，清理无用代码
