+++ src/session-types.ts
@@ -1,0 +1,162 @@
+import type { McpServerConfig, PermissionScope, PermissionSettings } from "./settings";
+import type { AskPermissionRequest, MessageToolPermission, UserToolPermission } from "./common/permissions";
+import type { CreateOpenAIClient } from "./tools/executor";
+import { DEEPSEEK_V4_MODELS } from "./common/model-capabilities";
+
+export type SessionStatus =
+  | "failed"
+  | "pending"
+  | "processing"
+  | "waiting_for_user"
+  | "completed"
+  | "interrupted"
+  | "ask_permission"
+  | "permission_denied";
+
+export type ModelUsage = {
+  prompt_tokens: number;
+  completion_tokens: number;
+  total_tokens: number;
+  completion_tokens_details?: Record<string, unknown>;
+  prompt_tokens_details?: Record<string, unknown>;
+  prompt_cache_hit_tokens?: number;
+  prompt_cache_miss_tokens?: number;
+  total_reqs?: number;
+};
+
+export type SessionProcessEntry = {
+  startTime: string;
+  command: string;
+  timeoutMs?: number;
+  deadlineAt?: string;
+  timedOut?: boolean;
+};
+
+export type BashTimeoutAdjustment = {
+  processId: string;
+  timeoutMs: number;
+  deadlineAt: string;
+  timedOut: boolean;
+};
+
+export type SessionEntry = {
+  id: string;
+  summary: string | null;
+  assistantReply: string | null;
+  assistantThinking: string | null;
+  assistantRefusal: string | null;
+  toolCalls: unknown[] | null;
+  status: SessionStatus;
+  failReason: string | null;
+  usage: ModelUsage | null;
+  usagePerModel: Record<string, ModelUsage> | null;
+  activeTokens: number;
+  createTime: string;
+  updateTime: string;
+  processes: Map<string, SessionProcessEntry> | null;
+  askPermissions?: AskPermissionRequest[];
+};
+
+export type SessionsIndex = {
+  version: 1;
+  entries: SessionEntry[];
+  originalPath: string;
+};
+
+export type SessionMessageRole = "system" | "user" | "assistant" | "tool";
+
+export type MessageMeta = {
+  function?: unknown;
+  paramsMd?: string;
+  resultMd?: string;
+  asThinking?: boolean;
+  isSummary?: boolean;
+  isModelChange?: boolean;
+  skill?: SkillInfo;
+  permissions?: MessageToolPermission[];
+  userPrompt?: UserPromptContent;
+};
+
+export type SessionMessage = {
+  id: string;
+  sessionId: string;
+  role: SessionMessageRole;
+  content: string | null;
+  contentParams: unknown | null;
+  messageParams: unknown | null;
+  compacted: boolean;
+  visible: boolean;
+  createTime: string;
+  updateTime: string;
+  meta?: MessageMeta;
+  html?: string;
+  checkpointHash?: string;
+};
+
+export type UndoTarget = {
+  message: SessionMessage;
+  index: number;
+  canRestoreCode: boolean;
+};
+
+export type UserPromptContent = {
+  text?: string;
+  imageUrls?: string[];
+  skills?: SkillInfo[];
+  permissions?: UserToolPermission[];
+  alwaysAllows?: PermissionScope[];
+};
+
+export type SkillInfo = {
+  name: string;
+  path: string;
+  description: string;
+  isLoaded?: boolean;
+};
+
+export type SessionManagerOptions = {
+  projectRoot: string;
+  createOpenAIClient: CreateOpenAIClient;
+  getResolvedSettings: () => {
+    model: string;
+    webSearchTool?: string;
+    mcpServers?: Record<string, McpServerConfig>;
+    permissions?: Required<PermissionSettings>;
+  };
+  renderMarkdown: (text: string) => string;
+  onAssistantMessage: (message: SessionMessage, shouldConnect: boolean) => void;
+  onSessionEntryUpdated?: (entry: SessionEntry) => void;
+  onLlmStreamProgress?: (progress: LlmStreamProgress) => void;
+  onMcpStatusChanged?: () => void;
+  onProcessStdout?: (pid: number, chunk: string) => void;
+};
+
+export type LlmStreamProgress = {
+  requestId: string;
+  sessionId?: string;
+  startedAt: string;
+  estimatedTokens: number;
+  formattedTokens: string;
+  phase: "start" | "update" | "end";
+};
+
+const DEFAULT_COMPACT_PROMPT_TOKEN_THRESHOLD = 128 * 1024;
+const DEEPSEEK_V4_COMPACT_PROMPT_TOKEN_THRESHOLD = 512 * 1024;
+
+export function getCompactPromptTokenThreshold(model: string): number {
+  return DEEPSEEK_V4_MODELS.has(model)
+    ? DEEPSEEK_V4_COMPACT_PROMPT_TOKEN_THRESHOLD
+    : DEFAULT_COMPACT_PROMPT_TOKEN_THRESHOLD;
+}
+
+function isUsageRecord(value: unknown): value is Record<string, unknown> {
+  return value !== null && typeof value === "object" && !Array.isArray(value);
+}
+
+export function getTotalTokens(usage: ModelUsage | null | undefined): number {
+  if (!isUsageRecord(usage)) {
+    return 0;
+  }
+  const totalTokens = (usage as Record<string, unknown>).total_tokens;
+  return typeof totalTokens === "number" ? totalTokens : 0;
+}
