+++ src/settings.ts
@@ -1,4 +1,7 @@
 import { defaultsToThinkingMode } from "./common/model-capabilities";
+import * as fs from "fs";
+import * as os from "os";
+import * as path from "path";
 
 export type DeepcodingEnv = Record<string, string | undefined> & {
   MODEL?: string;
@@ -369,4 +372,89 @@ export function applyModelConfigSelection(
   }
 
   return { settings: next, changed: true };
+}
+
+// ---------------------------------------------------------------------------
+// Default constants
+// ---------------------------------------------------------------------------
+
+export const DEFAULT_MODEL = "deepseek-v4-pro";
+export const DEFAULT_BASE_URL = "https://api.deepseek.com";
+
+// ---------------------------------------------------------------------------
+// Settings file I/O
+// ---------------------------------------------------------------------------
+
+export function getUserSettingsPath(): string {
+  return path.join(os.homedir(), ".deepcode", "settings.json");
+}
+
+export function getProjectSettingsPath(projectRoot: string): string {
+  return path.join(projectRoot, ".deepcode", "settings.json");
+}
+
+export function readSettingsFile(settingsPath: string): DeepcodingSettings | null {
+  try {
+    if (!fs.existsSync(settingsPath)) {
+      return null;
+    }
+    const raw = fs.readFileSync(settingsPath, "utf8");
+    return JSON.parse(raw) as DeepcodingSettings;
+  } catch {
+    return null;
+  }
+}
+
+export function readSettings(): DeepcodingSettings | null {
+  return readSettingsFile(getUserSettingsPath());
+}
+
+export function readProjectSettings(projectRoot: string = process.cwd()): DeepcodingSettings | null {
+  return readSettingsFile(getProjectSettingsPath(projectRoot));
+}
+
+function writeSettingsFile(settingsPath: string, settings: DeepcodingSettings): void {
+  fs.mkdirSync(path.dirname(settingsPath), { recursive: true });
+  fs.writeFileSync(settingsPath, `${JSON.stringify(settings, null, 2)}\n`, "utf8");
+}
+
+export function writeSettings(settings: DeepcodingSettings): void {
+  const settingsPath = getUserSettingsPath();
+  writeSettingsFile(settingsPath, settings);
+}
+
+export function writeProjectSettings(settings: DeepcodingSettings, projectRoot: string = process.cwd()): void {
+  const settingsPath = getProjectSettingsPath(projectRoot);
+  writeSettingsFile(settingsPath, settings);
+}
+
+export function writeModelConfigSelection(
+  selection: ModelConfigSelection,
+  current: ModelConfigSelection = resolveCurrentSettings(),
+  projectRoot: string = process.cwd()
+): { changed: boolean; settings: DeepcodingSettings } {
+  const projectSettingsPath = getProjectSettingsPath(projectRoot);
+  const shouldWriteProjectSettings = fs.existsSync(projectSettingsPath);
+  const rawSettings = shouldWriteProjectSettings ? readProjectSettings(projectRoot) : readSettings();
+  const result = applyModelConfigSelection(rawSettings, current, selection);
+  if (result.changed) {
+    if (shouldWriteProjectSettings) {
+      writeProjectSettings(result.settings, projectRoot);
+    } else {
+      writeSettings(result.settings);
+    }
+  }
+  return result;
+}
+
+export function resolveCurrentSettings(projectRoot: string = process.cwd()): ResolvedDeepcodingSettings {
+  return resolveSettingsSources(
+    readSettings(),
+    readProjectSettings(projectRoot),
+    {
+      model: DEFAULT_MODEL,
+      baseURL: DEFAULT_BASE_URL,
+    },
+    process.env
+  );
 }
