+++ src/tests/session.test.ts
@@ -5,7 +5,8 @@ import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
 import { GitFileHistory } from "../common/file-history";
-import { SessionManager, type SessionMessage } from "../session";
+import { type SessionMessage } from "../session-types";
+import { SessionManager } from "../session";
 
 const originalFetch = globalThis.fetch;
 const originalConsoleWarn = console.warn;
