+++ src/tests/thinkingState.test.ts
@@ -1,7 +1,7 @@
 import { test } from "node:test";
 import assert from "node:assert/strict";
 import { findExpandedThinkingId } from "../ui";
-import type { SessionMessage } from "../session";
+import type { SessionMessage } from "../session-types";
 
 function buildMessage(
   id: string,
