+++ src/ui/WelcomeScreen.tsx
@@ -2,11 +2,11 @@ import React, { useMemo, useState } from "react";
 import { Box, Text } from "ink";
 import * as os from "node:os";
 import path from "node:path";
-import type { SkillInfo } from "../session";
+import type { SkillInfo } from "../session-types";
 import type { ResolvedDeepcodingSettings } from "../settings";
 import { buildSlashCommands, BUILTIN_SLASH_COMMANDS, formatSlashCommandDescription } from "./slashCommands";
 import { ThemedGradient } from "./ThemedGradient";
-import { AsciiLogo } from "../AsciiArt";
+import { AsciiLogo } from "./AsciiArt";
 import { useAppContext } from "./contexts";
 
 type WelcomeScreenProps = {
