+++ src/ui/askUserQuestion.ts
@@ -1,4 +1,4 @@
-import type { SessionMessage, SessionStatus } from "../session";
+import type { SessionMessage, SessionStatus } from "../session-types";
 
 export type AskUserQuestionOption = {
   label: string;
