+++ src/ui/slashCommands.ts
@@ -1,4 +1,4 @@
-import type { SkillInfo } from "../session";
+import type { SkillInfo } from "../session-types";
 
 export type SlashCommandKind =
   | "skill"
